% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynpanel.R
\name{dpd.formula}
\alias{dpd.formula}
\title{formula}
\usage{
\method{dpd}{formula}(formula, data = list(), index = c("id", "time"), p,
  meth = c(0, 1, 2, 3, 4), ...)
}
\arguments{
\item{formula}{PIB~INF+TIR}

\item{data}{the dataframe}

\item{index}{: id is the name of the identity groups and time is the time per group}

\item{p}{scalar, autoregressive order for dependent variable}

\item{meth}{scalar, indicator for the Instruments to use}

\item{...}{not used}
}
\description{
formula
}

