% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynpanel.R
\name{dpd}
\alias{dpd}
\title{method}
\usage{
dpd(x, ...)
}
\arguments{
\item{x}{a numeric design matrix for the model.}

\item{...}{not used}
}
\description{
method
}
\author{
Zaghdoudi Taha
}

