% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynpanel.R
\docType{data}
\name{Produc}
\alias{Produc}
\title{US States Production}
\format{A data frame with 816 rows and 10 variables}
\usage{
data(Produc)
}
\description{
\itemize{
\item{state}{the state}
\item{year}{the year}
\item{pcap}{private capital stock}
\item{hwy}{highway and streets}
\item{water}{water and sewer facilities}
\item{util}{other public buildings and structures}
\item{pc}{public capital}
\item{gsp}{gross state products}
\item{emp}{labor input measured by the employement in non--agricultural payrolls}
\item{unemp}{state unemployment rate}
}
}

