% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynpanel.R
\name{summary.dpd}
\alias{summary.dpd}
\title{Summary}
\usage{
\method{summary}{dpd}(object, ...)
}
\arguments{
\item{object}{is the object of the function}

\item{...}{not used}
}
\description{
Summary
}

